+++ mlflow_faculty/__init__.py
@@ -13,7 +13,7 @@
 # limitations under the License.
 
 
-from mlflow_faculty.trackingstore import FacultyRestStore  # noqa F401
+from mlflow_faculty.tracking import FacultyRestStore  # noqa F401
 from mlflow_faculty.artifact_repository import (  # noqa F401
     FacultyDatasetsArtifactRepository,
 )
